Break out all test cases out of mod.rs.
