+++ semver_tests/src/test_cases/unit_struct_changed_kind.rs
@@ -1,0 +1,13 @@
+#[cfg(not(feature = "unit_struct_changed_kind"))]
+pub struct UnitStructToPlain;
+
+#[cfg(feature = "unit_struct_changed_kind")]
+pub struct UnitStructToPlain {}
+
+#[cfg(not(feature = "unit_struct_changed_kind"))]
+#[non_exhaustive]
+pub struct NonExhaustiveUnitStructToPlain;
+
+#[cfg(feature = "unit_struct_changed_kind")]
+#[non_exhaustive]
+pub struct NonExhaustiveUnitStructToPlain {}
