+++ src/test_data/enum_missing.output.ron
@@ -1,9 +1,14 @@
 [
     {
         "name": String("WillBeRemovedEnum"),
-        "path": List([String("semver_tests"), String("test_cases"), String("WillBeRemovedEnum")]),
+        "path": List([
+            String("semver_tests"),
+            String("test_cases"),
+            String("item_missing"),
+            String("WillBeRemovedEnum"),
+        ]),
         "visibility_limit": String("public"),
-        "span_filename": String("src/test_cases/mod.rs"),
-        "span_begin_line": Uint64(11),
+        "span_filename": String("src/test_cases/item_missing.rs"),
+        "span_begin_line": Uint64(7),
     }
 ]
