+++ src/test_data/enum_variant_missing.output.ron
@@ -2,7 +2,12 @@
     {
         "enum_name": String("VariantWillBeRemoved"),
         "variant_name": String("Bar"),
-        "path": List([String("semver_tests"), String("test_cases"), String("enum_variant_missing"), String("VariantWillBeRemoved")]),
+        "path": List([
+            String("semver_tests"),
+            String("test_cases"),
+            String("enum_variant_missing"),
+            String("VariantWillBeRemoved"),
+        ]),
         "visibility_limit": String("public"),
         "span_filename": String("src/test_cases/enum_variant_missing.rs"),
         "span_begin_line": Uint64(6),
