+++ src/test_data/struct_missing.output.ron
@@ -2,9 +2,14 @@
     {
         "name": String("WillBeRemovedStruct"),
         "struct_type": String("unit"),
-        "path": List([String("semver_tests"), String("test_cases"), String("WillBeRemovedStruct")]),
+        "path": List([
+            String("semver_tests"),
+            String("test_cases"),
+            String("item_missing"),
+            String("WillBeRemovedStruct"),
+        ]),
         "visibility_limit": String("public"),
-        "span_filename": String("src/test_cases/mod.rs"),
-        "span_begin_line": Uint64(7),
+        "span_filename": String("src/test_cases/item_missing.rs"),
+        "span_begin_line": Uint64(3),
     }
 ]
