+++ src/test_data/unit_struct_changed_kind.output.ron
@@ -1,9 +1,14 @@
 [
     {
         "name": String("UnitStructToPlain"),
-        "path": List([String("semver_tests"), String("test_cases"), String("UnitStructToPlain")]),
+        "path": List([
+            String("semver_tests"),
+            String("test_cases"),
+            String("unit_struct_changed_kind"),
+            String("UnitStructToPlain"),
+        ]),
         "visibility_limit": String("public"),
-        "span_filename": String("src/test_cases/mod.rs"),
-        "span_begin_line": Uint64(17),
+        "span_filename": String("src/test_cases/unit_struct_changed_kind.rs"),
+        "span_begin_line": Uint64(5),
     }
 ]
